pyproject.toml
[tool.poetry]
name = "sightlines"
version = "0.1.0"
description = ""
authors = ["Toby McLaughlin <toby@jarpy.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
launchpad-py = "^0.9.1"
pygame = "^2.1.2"
apscheduler = "^3.9.1.post1"
pdpyras = "^4.5.2"
python-dateutil = "^2.8.2"

[tool.poetry.scripts]
sightlines = "sightlines.main:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
flake8 = "^6.0.0"
types-requests = "^2.28.11.7"
types-python-dateutil = "^2.8.19.5"
ipython = "^8.8.0"

[tool.mypy]
python_version = "3.10"
explicit_package_bases = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


